feat(#21): reorganize unit tests
