--- core/src/dsnp/tests/mod.rs
@@ -1,5 +1,0 @@
-mod compression;
-mod dsnp_types;
-mod encryption;
-mod pseudo_relationship_identifier;
-mod schema;
