+++ core/src/graph/mod.rs
@@ -1,7 +1,5 @@
 pub mod graph;
 mod key_manager;
 pub mod page;
-#[cfg(test)]
-mod test_helpers;
 pub mod updates;
 pub mod user;
