+++ core/src/lib.rs
@@ -2,5 +2,7 @@ pub mod api;
 mod dsnp;
 mod frequency;
 mod graph;
+#[cfg(test)]
+mod tests;
 mod types;
 mod util;
